Zelix string transformer
